requirements.txt
fastapi
uvicorn[standard]
pandas
pyarrow
groq
python-dotenv
pydantic
cachetools
streamlit
